Update PresentationsController.php
This is fine as long as the PresentationTypes start at ID one, and are consecutive.  What happens when/if a PresentationType gets deleted, then another one added?  The auto-increment index doesn't get reset on deletion, so the new PresentationType record will get the next auto-increment value.  In this situation, you could have PresentationType records with ids 1-9, and 11.  The position in the array will no longer match the ID.

I suggest not using the array index as a substitute for the ID, and just pull the ID out of the PresentationType object in your view.