Notes on RolesController.php
**findOrFail()**
See notes on PresentationController.php

**index()**
Shouldn't access to this method be restricted?  A list of all users and their roles is definitely privileged information.  Maybe it is and I just don't understand Laravel.  

** new_role() and approve() **
It's not clear from the method names that these two methods turn someone into a professor.  I would recommend renaming these to: `requestProfessorRole()` and `approveProfessorRole()`

**modify()** should handle `save()` failing.